Make the keyboard configuration optional
It interferes quite a lot with Grafana
